Add files via upload
Welcome to Coinswitch!

Description:

Coinswitch is a game that involves two coins (red and blue) constantly rotating around a tile in a pre-defined path. The user has to hit the space bar to land the coin on a particular tile on the path, failing to do so, the user loses the game. If the user manages to land all coins onto the tile and reaches the end point, he wins! The path is a maze in the medium and the hard modes, and the user must figure out the path to the endpoint. These modes also have special tiles, that are challenges to the user's path. 

How to Run:

Hit command B or the Python Run button, and the splash screen for the game will open. Click anywhere on the screen, and enter your name. The instructions are also displayed on the game, go ahead and read them once more, and then move on to play the easy, medium, or hard modes. If you lose your turn, you come back to the main menu. There is a leaderboard to see players and their respective high scores in the easy mode. The leaderboard only displays the top 10

Libraries: 

Pygame for sound

Shortcut commands:

Space bar: Playing
r: to go back to the main menu